+++ src/handlers/index.ts
@@ -4,6 +4,7 @@ require('./hello')
 require('./respond')
 require('./moderate')
 require('./command')
+require('./no-duplicate')
 
 export * from './abstract'
 export * from './quotes'
@@ -11,3 +12,4 @@ export * from './hello'
 export * from './respond'
 export * from './moderate'
 export * from './command'
+export * from './no-duplicate'